+++ packages/i18n/src/useNextIntl.tsx
@@ -1,0 +1,5 @@
+import { useTranslations } from 'next-intl'
+
+import { useIntlImpl } from './useIntlImpl'
+
+export const useNextIntl = () => useIntlImpl(useTranslations)
