+++ packages/i18n/tsconfig.json
@@ -4,5 +4,6 @@
     "lib": ["esnext", "dom"],
     "outDir": "./lib",
     "rootDir": "./src"
-  }
+  },
+  "exclude": ["./src/**/*.test.*"]
 }
